Moving BadgedImageView to a more appropriate location
